services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "beyul-api"
version = "0.1.0"
description = "FastAPI service for Beyul"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0,<0.117.0",
  "uvicorn[standard]>=0.35.0,<0.36.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "redis>=5.2.0,<6.0.0",
  "sqlalchemy>=2.0.38,<3.0.0",
  "asyncpg>=0.30.0,<0.31.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<0.29.0",
  "pytest>=8.3.0,<9.0.0"
]

[tool.setuptools]

[tool.setuptools.packages.find]
include = ["app*"]
